+++ packages/components-react/projects/nextjs/components/FooterPartials.tsx
@@ -1,5 +1,0 @@
-import { getDSRPonyfill } from '@porsche-design-system/components-react/partials';
-
-export const FooterPartials = (): JSX.Element => {
-  return <>{getDSRPonyfill({ format: 'jsx' })}</>;
-};