+++ packages/components-react/projects/nextjs/components/index.ts
@@ -1,4 +1,3 @@
-export * from './FooterPartials';
 export * from './HeaderPartials';
 export * from './TableWithCaption';
 export * from './Toast';