+++ packages/components-react/projects/remix/app/body-partials.server.tsx
@@ -1,5 +1,0 @@
-import { getDSRPonyfill } from '@porsche-design-system/components-react/partials';
-
-export const BodyPartials = (): JSX.Element => {
-  return <>{getDSRPonyfill({ format: 'jsx' })}</>;
-};