+++ packages/storefront/src/models/partials.ts
@@ -1,6 +1,5 @@
 export type Partials =
   | 'getComponentChunkLinks'
-  | 'getDSRPonyfill'
   | 'getFontFaceStyles'
   | 'getFontFaceStylesheet'
   | 'getFontLinks'