+++ packages/storefront/src/utils/partials/getNextPartialExample.ts
@@ -7,23 +7,6 @@ export const getNextPartialExample = (name: Partials) => {
 const nextExamples: {
   [partial in Partials]: string;
 } = {
-  getDSRPonyfill: `/* ./app/layout.tsx */
-import React from 'react';
-import { preload } from 'react-dom';
-import { getDSRPonyfill } from '@porsche-design-system/components-react/partials';
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <>
-      {children}
-      {getDSRPonyfill({ format: 'jsx' })}
-    </>
-  );
-}`,
   getMetaTagsAndIconLinks: `/* ./app/layout.tsx */
 import type { Metadata, Viewport } from "next";
 import { getMetaTagsAndIconLinks } from '@porsche-design-system/components-react/partials';